requirements.txt
python-telegram-bot==22.5
moviepy==2.2.1
Pillow==10.3.0

